+++ nextjs-blog/components/layout.tsx
@@ -8,8 +8,13 @@ import styles from "./layout.module.css";
 const name = "Your Name";
 export const siteTitle = "Next.js Sample Website";
 
-// eslint-disable-next-line react/prop-types
-export default function Layout({ children, home }) {
+export default function Layout({
+  children,
+  home,
+}: {
+  children: React.ReactNode;
+  home?: boolean;
+}) {
   return (
     <div className={styles.container}>
       <Head>
