+++ nextjs-blog/lib/posts.ts
@@ -21,7 +21,7 @@ export function getSortedPostsData() {
     // Combine the data with the id
     return {
       id,
-      ...matterResult.data,
+      ...(matterResult.data as { date: string }),
     };
   });
   // Sort posts by date
@@ -46,7 +46,7 @@ export function getAllPostIds() {
   });
 }
 
-export function getPostData(id) {
+export function getPostData(id: string) {
   const fullPath = path.join(postsDirectory, `${id}.md`);
   const fileContents = fs.readFileSync(fullPath, "utf8");
