--- nextjs-blog/pages/_app.js
@@ -1,7 +1,0 @@
-import React from "react";
-import "../styles/global.css";
-
-// eslint-disable-next-line react/prop-types
-export default function App({ Component, pageProps }) {
-  return <Component {...pageProps} />;
-}
