+++ nextjs-blog/pages/_app.tsx
@@ -1,0 +1,6 @@
+import { AppProps } from "next/app";
+import "../styles/global.css";
+
+export default function App({ Component, pageProps }: AppProps) {
+  return <Component {...pageProps} />;
+}
