+++ nextjs-blog/pages/posts/[id].tsx
@@ -1,0 +1,38 @@
+/* eslint-disable react/react-in-jsx-scope */
+import { GetStaticPaths, GetStaticProps } from "next";
+import Layout from "../../components/layout";
+import { getAllPostIds, getPostData } from "../../lib/posts";
+
+export default function Post({
+  postData,
+}: {
+  postData: { title: string; id: string; date: string };
+}) {
+  return (
+    <Layout>
+      {" "}
+      {postData.title}
+      <br />
+      {postData.id}
+      <br />
+      {postData.date}
+    </Layout>
+  );
+}
+
+export const getStaticPaths: GetStaticPaths = async () => {
+  const paths = getAllPostIds();
+  return {
+    paths,
+    fallback: false,
+  };
+};
+
+export const getStaticProps: GetStaticProps = async ({ params }) => {
+  const postData = await getPostData(params?.id as string);
+  return {
+    props: {
+      postData,
+    },
+  };
+};
